life-discovery/services/discovery_engine/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
playwright==1.49.1
python-dateutil==2.9.0.post0
